feat update: useCountries hook created for filter and search
